+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/feign/plugin/reporter/ExceptionPolarisReporter.java
@@ -18,7 +18,9 @@
 package com.tencent.cloud.rpc.enhancement.feign.plugin.reporter;
 
 import java.net.SocketTimeoutException;
+import java.util.ArrayList;
 
+import com.tencent.cloud.rpc.enhancement.AbstractPolarisReporterAdapter;
 import com.tencent.cloud.rpc.enhancement.config.RpcEnhancementReporterProperties;
 import com.tencent.cloud.rpc.enhancement.feign.plugin.EnhancedFeignContext;
 import com.tencent.cloud.rpc.enhancement.feign.plugin.EnhancedFeignPlugin;
@@ -26,28 +28,36 @@ import com.tencent.cloud.rpc.enhancement.feign.plugin.EnhancedFeignPluginType;
 import com.tencent.polaris.api.core.ConsumerAPI;
 import com.tencent.polaris.api.pojo.RetStatus;
 import com.tencent.polaris.api.rpc.ServiceCallResult;
+import com.tencent.polaris.client.api.SDKContext;
 import feign.Request;
 import feign.Response;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.springframework.core.Ordered;
+import org.springframework.http.HttpHeaders;
 
 /**
  * Polaris reporter when feign call fails.
  *
  * @author Haotian Zhang
  */
-public class ExceptionPolarisReporter implements EnhancedFeignPlugin {
+public class ExceptionPolarisReporter extends AbstractPolarisReporterAdapter implements EnhancedFeignPlugin {
 
 	private static final Logger LOG = LoggerFactory.getLogger(ExceptionPolarisReporter.class);
 	private final RpcEnhancementReporterProperties reporterProperties;
 
 	private final ConsumerAPI consumerAPI;
 
+	private final SDKContext context;
+
+
 	public ExceptionPolarisReporter(RpcEnhancementReporterProperties reporterProperties,
+			SDKContext context,
 			ConsumerAPI consumerAPI) {
+		super(reporterProperties);
 		this.reporterProperties = reporterProperties;
+		this.context = context;
 		this.consumerAPI = consumerAPI;
 	}
 
@@ -78,7 +88,13 @@ public class ExceptionPolarisReporter implements EnhancedFeignPlugin {
 			}
 			LOG.debug("Will report result of {}. Request=[{} {}]. Response=[{}]. Delay=[{}]ms.", retStatus.name(), request.httpMethod()
 					.name(), request.url(), response.status(), delay);
-			ServiceCallResult resultRequest = ReporterUtils.createServiceCallResult(request, response, delay, retStatus);
+			ServiceCallResult resultRequest = ReporterUtils.createServiceCallResult(this.context, request, response,
+					delay, retStatus, serviceCallResult -> {
+						HttpHeaders headers = new HttpHeaders();
+						response.headers().forEach((s, strings) -> headers.addAll(s, new ArrayList<>(strings)));
+						serviceCallResult.setRetStatus(getRetStatusFromRequest(headers, serviceCallResult.getRetStatus()));
+						serviceCallResult.setRuleName(getActiveRuleNameFromRequest(headers));
+					});
 			consumerAPI.updateServiceCallResult(resultRequest);
 		}
 	}